+++ src/Provider/Norway/Norway.php
@@ -1,0 +1,50 @@
+<?php
+
+/*
+ * This file is part of the umulmrum/holiday package.
+ *
+ * (c) Stefan Kruppa
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Umulmrum\Holiday\Provider\Norway;
+
+use Umulmrum\Holiday\Constant\HolidayName;
+use Umulmrum\Holiday\Constant\HolidayType;
+use Umulmrum\Holiday\Model\Holiday;
+use Umulmrum\Holiday\Model\HolidayList;
+use Umulmrum\Holiday\Provider\CommonHolidaysTrait;
+use Umulmrum\Holiday\Provider\HolidayProviderInterface;
+use Umulmrum\Holiday\Provider\Religion\ChristianHolidaysTrait;
+
+class Norway implements HolidayProviderInterface
+{
+    use ChristianHolidaysTrait;
+    use CommonHolidaysTrait;
+
+    public function calculateHolidaysForYear(int $year): HolidayList
+    {
+        $holidays = new HolidayList();
+        $holidays->add($this->getNewYear($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getMaundyThursday($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getGoodFriday($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF | HolidayType::PARTIAL_ONLY));
+        $holidays->add($this->getEasterSunday($year));
+        $holidays->add($this->getEasterMonday($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getLaborDay($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getConstitutionDay($year));
+        $holidays->add($this->getAscension($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getWhitSunday($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getWhitMonday($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getChristmasDay($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        $holidays->add($this->getSecondChristmasDay($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+
+        return $holidays;
+    }
+
+    protected function getConstitutionDay(int $year): Holiday
+    {
+        return Holiday::create(HolidayName::NORWAY_CONSTITUTION_DAY, "{$year}-05-17", HolidayType::OFFICIAL | HolidayType::DAY_OFF);
+    }
+}