+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Holidays for Norway
+- Holidays for Turkey
+
 ## [0.7.0] - 2024-05-25
 
 ### Added